iUpdated app with functionality and styling
Added a tweet container to hold all submitted tweets.

Added a hover state to tweets.

Added functionality to dynamically render the tweet data.Used jQuery in generating a structure for submitted tweets.

Used jQuery and AJAX to enable submission of tweets asynchronously.

Used AJAX to enable fetching of data from the server.

Added validation to the new tweet form to disallow empty tweets and tweets containing over 140 characters.

Enabled fetching of tweets on submission of new tweet form so users can now see their submitted tweet loaded below the form.

Added an escape function to prevent cross-site scripting based on user inputs.

Added a button to toggle the view of the new tweet form. It will hide and reveal the new tweet form based on selection of the "Compose" button. The new tweet field will be in focus upon activation of the new tweet form.

Created error messages to complement the new tweet form validation rules and display to the user the actions to take to successfully submit their tweet.